frontend/requirements.txt
streamlit
requests
pillow
opencv-python
numpy
facenet-pytorch
matplotlib
facenet-pytorch
lightgbm
shap
scipy